+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][keepachangelog], and this project
 adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
+### Added
+* Add the `from_state` method to start the machine from any given state.
 
 ## [0.3.0] - 2019-05-22
 ### Changed